requirements.txt
# Основные зависимости для Android приложения
kivy>=2.1.0
kivymd>=1.1.1

# Компьютерное зрение
opencv-python>=4.8.0
numpy>=1.24.0

# Работа с USB/OTG (только для разработки)
pyusb>=1.2.1

# Дополнительные утилиты
plyer>=2.1.0
python-for-android>=2023.10.06

# Логирование и отладка
colorlog>=6.7.0

# Примечание: android>=0.4 устанавливается автоматически при сборке APK